refactor: CarFilter API를 Filter API에서 처리하도록 변경 (양방향 의존에서 단방향 의존으로 변경)
