+++ backend/src/docs/asciidoc/car.adoc
@@ -40,29 +40,3 @@ include::{snippets}/car-controller-test/delete_car/http-request.adoc[]
 === Response
 
 include::{snippets}/car-controller-test/delete_car/http-response.adoc[]
-
-== 차량에 등록된 필터를 조회한다 (/cars/{carId}/filters)
-
-=== Request
-
-include::{snippets}/car-controller-test/find_car_filters/path-parameters.adoc[]
-include::{snippets}/car-controller-test/find_car_filters/http-request.adoc[]
-
-=== Response
-
-include::{snippets}/car-controller-test/find_car_filters/response-fields.adoc[]
-include::{snippets}/car-controller-test/find_car_filters/http-response.adoc[]
-
-== 차량에 필터를 등록한다 (/cars/{carId}/filters)
-
-=== Request
-
-include::{snippets}/car-controller-test/add_car_filters/request-headers.adoc[]
-include::{snippets}/car-controller-test/add_car_filters/path-parameters.adoc[]
-include::{snippets}/car-controller-test/add_car_filters/request-body.adoc[]
-include::{snippets}/car-controller-test/add_car_filters/http-request.adoc[]
-
-=== Response
-
-include::{snippets}/car-controller-test/add_car_filters/response-fields.adoc[]
-include::{snippets}/car-controller-test/add_car_filters/http-response.adoc[]
