+++ backend/src/test/java/com/carffeine/carffeine/car/controller/CarControllerTest.java
@@ -5,11 +5,6 @@ import com.carffeine.carffeine.car.infrastructure.dto.CarResponse;
 import com.carffeine.carffeine.car.infrastructure.dto.CarsResponse;
 import com.carffeine.carffeine.car.service.dto.CarRequest;
 import com.carffeine.carffeine.car.service.dto.CarsRequest;
-import com.carffeine.carffeine.filter.controller.dto.FiltersResponse;
-import com.carffeine.carffeine.filter.domain.Filter;
-import com.carffeine.carffeine.filter.domain.FilterType;
-import com.carffeine.carffeine.filter.service.dto.FilterRequest;
-import com.carffeine.carffeine.filter.service.dto.FiltersRequest;
 import com.carffeine.carffeine.helper.MockBeanInjection;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.http.HttpHeaders;
@@ -27,12 +22,8 @@ import java.util.List;
 
 import static com.carffeine.carffeine.car.fixture.CarFixture.createCar;
 import static com.carffeine.carffeine.car.fixture.CarFixture.createOtherCar;
-import static com.carffeine.carffeine.filter.fixture.FilterFixture.createCapacityFilter;
-import static com.carffeine.carffeine.filter.fixture.FilterFixture.createCompanyFilter;
-import static com.carffeine.carffeine.filter.fixture.FilterFixture.createConnectorTypeFilter;
 import static com.carffeine.carffeine.helper.RestDocsHelper.customDocument;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.when;
@@ -141,78 +132,6 @@ class CarControllerTest extends MockBeanInjection {
                 .andDo(customDocument("delete_car",
                         requestHeaders(headerWithName(HttpHeaders.AUTHORIZATION).description("인증 토큰")),
                         pathParameters(parameterWithName("carId").description("차량 id"))
-                ));
-    }
-
-    @Test
-    void 차량_필터를_모두_조회한다() throws Exception {
-        // given
-        List<Filter> filters = List.of(
-                createCompanyFilter(),
-                createConnectorTypeFilter(),
-                createCapacityFilter()
-        );
-        FiltersResponse response = FiltersResponse.from(filters);
-
-        // when
-        when(filterQueryService.findCarFilters(anyLong())).thenReturn(response);
-
-        // then
-        mockMvc.perform(get("/cars/{carId}/filters", 1L))
-                .andExpect(status().isOk())
-                .andDo(customDocument("find_car_filters",
-                        pathParameters(parameterWithName("carId").description("차량 id")),
-                        responseFields(
-                                fieldWithPath("companies[0]").type(JsonFieldType.ARRAY).description("충전기 회사"),
-                                fieldWithPath("capacities[0]").type(JsonFieldType.ARRAY).description("충전 용량"),
-                                fieldWithPath("connectorTypes[0]").type(JsonFieldType.ARRAY).description("충전기 타입")
-                        )
-                ));
-    }
-
-    @Test
-    void 차량_필터를_저장한다() throws Exception {
-        // given
-        List<Filter> filters = List.of(
-                createCompanyFilter(),
-                createConnectorTypeFilter(),
-                createCapacityFilter()
-        );
-
-        FiltersRequest request = new FiltersRequest(
-                List.of(
-                        new FilterRequest(FilterType.COMPANY.getName(), "HG"),
-                        new FilterRequest(FilterType.CAPACITY.getName(), "2.00"),
-                        new FilterRequest(FilterType.CONNECTOR_TYPE.getName(), "DC_COMBO")
-                )
-        );
-
-        // when
-        when(carService.addCarFilters(anyLong(), anyLong(), any())).thenReturn(filters);
-
-        // then
-        mockMvc.perform(post("/cars/{carId}/filters", 1L)
-                        .header(AUTHORIZATION, "Bearer token~~")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(request))
-                )
-                .andExpect(status().isCreated())
-                .andDo(customDocument("add_car_filters",
-                        requestHeaders(headerWithName(HttpHeaders.AUTHORIZATION).description("인증 토큰")),
-                        pathParameters(parameterWithName("carId").description("차량 id")),
-                        requestFields(
-                                fieldWithPath("filters[0].type").type(JsonFieldType.STRING).description("필터 종류"),
-                                fieldWithPath("filters[0].name").type(JsonFieldType.STRING).description("필터 이름"),
-                                fieldWithPath("filters[1].type").type(JsonFieldType.STRING).description("필터 종류"),
-                                fieldWithPath("filters[1].name").type(JsonFieldType.STRING).description("필터 이름"),
-                                fieldWithPath("filters[2].type").type(JsonFieldType.STRING).description("필터 종류"),
-                                fieldWithPath("filters[2].name").type(JsonFieldType.STRING).description("필터 이름")
-                        ),
-                        responseFields(
-                                fieldWithPath("companies[0]").type(JsonFieldType.ARRAY).description("충전기 회사"),
-                                fieldWithPath("capacities[0]").type(JsonFieldType.ARRAY).description("충전 용량"),
-                                fieldWithPath("connectorTypes[0]").type(JsonFieldType.ARRAY).description("충전기 타입")
-                        )
                 ));
     }
 }
