+++ backend/src/test/java/com/carffeine/carffeine/filter/service/FilterServiceTest.java
@@ -1,5 +1,12 @@
 package com.carffeine.carffeine.filter.service;
 
+import com.carffeine.carffeine.admin.exception.AdminException;
+import com.carffeine.carffeine.admin.exception.AdminExceptionType;
+import com.carffeine.carffeine.car.domain.Car;
+import com.carffeine.carffeine.car.domain.CarFilterRepository;
+import com.carffeine.carffeine.car.domain.CarRepository;
+import com.carffeine.carffeine.car.fake.FakeCarFilterRepository;
+import com.carffeine.carffeine.car.fake.FakeCarRepository;
 import com.carffeine.carffeine.filter.domain.Filter;
 import com.carffeine.carffeine.filter.domain.FilterRepository;
 import com.carffeine.carffeine.filter.domain.FilterType;
@@ -17,7 +24,9 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
+import static com.carffeine.carffeine.car.fixture.CarFixture.createCar;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
@@ -27,17 +36,26 @@ class FilterServiceTest {
     private FilterService filterService;
     private FilterRepository filterRepository;
     private MemberRepository memberRepository;
+    private CarRepository carRepository;
+    private CarFilterRepository carFilterRepository;
 
     private Member admin;
+    private Member member;
 
     @BeforeEach
     void setup() {
         filterRepository = new FakeFilterRepository();
         memberRepository = new FakeMemberRepository();
-        filterService = new FilterService(filterRepository, memberRepository);
+        carRepository = new FakeCarRepository();
+        carFilterRepository = new FakeCarFilterRepository();
+        filterService = new FilterService(filterRepository, memberRepository, carRepository, carFilterRepository);
 
         admin = memberRepository.save(Member.builder()
                 .memberRole(MemberRole.ADMIN)
+                .build());
+
+        member = memberRepository.save(Member.builder()
+                .memberRole(MemberRole.USER)
                 .build());
     }
 
@@ -73,7 +91,7 @@ class FilterServiceTest {
 
         // then
         List<Filter> filters = filterService.findAllFilters();
-        assertThat(filters.size()).isEqualTo(3);
+        assertThat(filters).hasSize(3);
     }
 
     @Test
@@ -94,6 +112,40 @@ class FilterServiceTest {
 
         // then
         List<Filter> filters = filterService.findAllFilters();
-        assertThat(filters.size()).isEqualTo(2);
+        assertThat(filters).hasSize(2);
+    }
+
+    @Test
+    void 차량에_필터를_적용한다() {
+        // given
+        Car savedCar = carRepository.save(createCar());
+        filterRepository.saveAll(List.of(new Filter("HG", FilterType.COMPANY)));
+        String filterName = "HG";
+        FiltersRequest request = new FiltersRequest(
+                List.of(new FilterRequest(FilterType.COMPANY.getName(), filterName))
+        );
+
+        // when
+        List<Filter> result = filterService.addCarFilters(admin.getId(), savedCar.getId(), request);
+
+        // then
+        assertSoftly(softly -> {
+            softly.assertThat(result.size()).isEqualTo(1);
+            softly.assertThat(result.get(0).getName()).isEqualTo(filterName);
+        });
+    }
+
+    @Test
+    void 어드민이_아니면_차량에_필터를_적용하지_못한다() {
+        // given
+        Car savedCar = carRepository.save(createCar());
+        FiltersRequest request = new FiltersRequest(
+                List.of(new FilterRequest(FilterType.COMPANY.getName(), "HG"))
+        );
+
+        // when & then
+        assertThatThrownBy(() -> filterService.addCarFilters(member.getId(), savedCar.getId(), request))
+                .isInstanceOf(AdminException.class)
+                .hasMessage(AdminExceptionType.NOT_ADMIN.message());
     }
 }
